Split geometry and generic
