+++ structuralcodes/materials/__init__.py
@@ -3,7 +3,4 @@ from . import concrete
 from . import constitutive_laws
 
 
-__all__ = [
-    'concrete',
-    'constitutive_laws'
-]
+__all__ = ['concrete', 'constitutive_laws']
